+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.9.3] - 2023-06-XX
 
+### Fixed
+- Fixed a bug that was preventing copying correct S3 file & folder paths in the File Manager.
+
 ### Changed
 - Updated and simplified the data model for storage drives and folders to make them easier to work with in the long-term.