+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fixed a bug that was preventing copying correct S3 file & folder paths in the File Manager.
 - Fixed bug preventing adding folders to assays. 
-- Fixed file manager breadcrumb traversal.
+- Fixed file manager traversal via controls & breadcrumbs.
 
 ### Changed
 - Updated and simplified the data model for storage drives and folders to make them easier to work with in the long-term.