+++ CHANGELOG.md
@@ -9,7 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Working on a new module for capturing data sets and their metadata. This module will allow the registration of various data sources (such as AWS S3), and the creation of data set records that will include one-or-more resources from these storage locations, along with metadata describing them. Users will then be allowed to associate one-or-more data sets with a study. Studies will likely get data set records created for them, which will include their storage folder and notebook by default.
 - Will add a notifications feature that will capture and display notifications for users.
 
-## [0.9.3] - 2023-06-XX
+## [0.9.4] - 2023-09-XX
+
+### Changed
+- GitLab repositories now get more specific paths created for them, including both the project codes and names. GitLab repositories also now get a better truncated description based on the study/assay description.
+
+## [0.9.3] - 2023-07-04
 
 ### Added
 - You can now create S3 folders for assays in the same way you can for studies.